# microservices-with-restful/restful-api/docs/swagger.yaml
basePath: /api
definitions:
  model.Identifier:
    properties:
      name:
        type: string
      value:
        type: number
    type: object
  model.Plc:
    properties:
      identifier:
        items:
          $ref: '#/definitions/model.Identifier'
        type: array
      machine:
        type: string
      time:
        type: string
    type: object
  model.State:
    properties:
      color:
        type: string
      machine:
        type: string
      name:
        type: string
      time:
        type: string
      value:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a RESTful API for a plc and state data.
  title: RESTful API
  version: "1.0"
paths:
  /plcs/{machine}/{time}/{limit}:
    get:
      description: Get plc data
      parameters:
      - description: Machine
        in: path
        name: machine
        required: true
        type: string
      - description: Time
        in: path
        name: time
        required: true
        type: string
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Plc'
            type: array
        "500":
          description: error
          schema:
            type: string
      summary: Get plc data
      tags:
      - plc
  /states:
    post:
      consumes:
      - application/json
      description: Create states
      parameters:
      - description: States
        in: body
        name: states
        required: true
        schema:
          items:
            $ref: '#/definitions/model.State'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Create states
      tags:
      - state
  /states/{machine}/{limit}:
    get:
      description: Get state data
      parameters:
      - description: Machine
        in: path
        name: machine
        required: true
        type: string
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.State'
            type: array
        "500":
          description: error
          schema:
            type: string
      summary: Get state data
      tags:
      - state
swagger: "2.0"